refactor(gogetutils): extract to gogetutils
